requirements.txt
tensorflow>=2.6.0
pillow>=8.0.0
numpy>=1.19.2
scikit-learn>=0.24.0
optuna>=2.10.0
pandas>=1.3.0
matplotlib>=3.4.0